WIP: CLI Cadastrar prestador, config e cliente funcionando

+++ PyNFSe/cli/file_manipulation.py
@@ -1,25 +1,30 @@
 import json
-import pickle
+import os
 
-def file_create(filename):
-    with open(filename, mode='w') as file:
-        json.dump({}, file)
 
-def append_json(filename, value):
-    with open(filename, mode='r') as file:
-        json_file = json.load(file)
+def home_folder():
+    return os.path.expanduser('~')
 
-    with open(filename, mode='w') as file:
-        json_file.append(value)
-        json.dump(json_file, file, ensure_ascii=False)
 
-def add_client(clients_file, client_dict):
-    with open(clients_file, mode='r') as file:
-        json_file = json.load(file)
+def criar_diretorio(directory):
+    path = os.path.join(home_folder(), directory)
 
-    index = client_dict['numero_documento']
-    client_json = json.dumps(client_dict, ensure_ascii=False)
+    try:
+        os.mkdir(path)
+        print('Diretório {} criado'.format(directory))
+    except FileExistsError:
+        print('O diretório {} já existe'.format(directory))
 
-    with open(clients_file, mode='w') as file:
-        json_file[index] = (client_json)
-        json.dump(json_file, file, ensure_ascii=False)
+
+def criar_arquivo_json(filename):
+    # path = os.path.join(home_folder(), directory)
+    # file_path = os.path.join(path, filename)
+
+    if not os.path.exists(filename):
+        with open(filename, mode='w') as file:
+            json.dump({}, file)
+        print('Arquivo {} criado'.format(filename))
+        return True
+    else:
+        print('O arquivo {} já existe'.format(filename))
+        return False
